backend/requirements.txt
Django>=4.2,<5.0
psycopg[binary]>=3.1
python-dotenv>=1.0
djangorestframework>=3.15
django-cors-headers>=4.4
PyJWT>=2.8
google-auth>=2.35
requests>=2.32
